chore : component 분리
