+++ src/pages/index.js
@@ -1,3 +1,2 @@
 export { default as Home } from 'pages/home';
 export { default as Mypage } from 'pages/Mypage/Mypage';
-export { default as MypageEdit } from 'pages/Mypage/MypageEdit';
